+++ bqueryd/node.py
@@ -8,34 +8,35 @@ import bqueryd
 
 config = configobj.ConfigObj('/etc/bqueryd.cfg')
 redis_url = config.get('redis_url', 'redis://127.0.0.1:6379/0')
+azure_conn_string = config.get('azure_conn_string', None)
 
 
-def main():
-    if '-vvv' in sys.argv:
+def main(argv=sys.argv):
+    if '-vvv' in argv:
         loglevel = logging.DEBUG
-    elif '-vv' in sys.argv:
+    elif '-vv' in argv:
         loglevel = logging.INFO
-    elif '-v' in sys.argv:
+    elif '-v' in argv:
         loglevel = logging.WARNING
     else:
         loglevel = logging.ERROR
 
     data_dir = bqueryd.DEFAULT_DATA_DIR
-    for arg in sys.argv:
+    for arg in argv:
         if arg.startswith('--data_dir='):
             data_dir = arg[11:]
 
-    if 'controller' in sys.argv:
-        bqueryd.ControllerNode(redis_url=redis_url, loglevel=loglevel).go()
-    elif 'worker' in sys.argv:
+    if 'controller' in argv:
+        bqueryd.ControllerNode(redis_url=redis_url, loglevel=loglevel, azure_conn_string=azure_conn_string).go()
+    elif 'worker' in argv:
         bqueryd.WorkerNode(redis_url=redis_url, loglevel=loglevel, data_dir=data_dir).go()
-    elif 'downloader' in sys.argv:
-        bqueryd.DownloaderNode(redis_url=redis_url, loglevel=loglevel).go()
-    elif 'movebcolz' in sys.argv:
+    elif 'downloader' in argv:
+        bqueryd.DownloaderNode(redis_url=redis_url, loglevel=loglevel, azure_conn_string=azure_conn_string).go()
+    elif 'movebcolz' in argv:
         bqueryd.MoveBcolzNode(redis_url=redis_url, loglevel=loglevel).go()
     else:
-        if len(sys.argv) > 1 and sys.argv[1].startswith('tcp:'):
-            rpc = bqueryd.RPC(address=sys.argv[1], redis_url=redis_url, loglevel=loglevel)
+        if len(argv) > 1 and argv[1].startswith('tcp:'):
+            rpc = bqueryd.RPC(address=argv[1], redis_url=redis_url, loglevel=loglevel)
         else:
             rpc = bqueryd.RPC(redis_url=redis_url, loglevel=loglevel)
         import IPython