+++ RELEASE_NOTES.rst
@@ -1,3 +1,39 @@
+Release 0.3.8
+=======================
+- Fix download error with unzipping ubication.
+
+Release 0.3.7
+=======================
+- Update minimum requirements.
+
+Release 0.3.6
+=======================
+- Ensure uncompressing temporary files are deleted if there is an error uncompressing.
+
+Release 0.3.5
+=======================
+- Added functionality for downloading files from Azure Blob Storage.
+
+Release 0.3.4
+=======================
+- Fix module import error.
+
+Release 0.3.3
+=======================
+- Fix hardcoded version string in ``bqueryd.__init__``.
+- Round robin over all available workers (both local and remote) in the controller unless "needs_local" is specified.
+
+Release 0.3.2
+=======================
+- Fix the issue with docker container exiting when run through docker-compose.
+
+Release 0.3.1
+=======================
+- Add docker-compose to containerize the project.
+- Add circleci config.
+- Add tests for downloader, rpc methods and movebcolz.
+- Simplify node restart check.
+
 Release  0.3.0
 =======================
 - Move to boto3